MLBEDSW-9889: MLCE: RSQRT int16 produces output diff
 - Update LUT table generation for RSQRT int16 to use
float to match reference and avoid rounding issues

+++ ethosu/regor/compiler/tflite_graph_optimiser.cpp
@@ -2575,7 +2575,7 @@ Operation *TFLiteGraphOptimiser::ConvertRSqrtToLUT(Graph *const graph, Operation
         const auto ofmScale = operation->Output(TensorUsage::OFM)->quantization.scales[0].Dequantize();
         returnOp = ConvertToInterpolatingLUT16(
             operation,
-            [&ofmScale](double x) -> double
+            [&ofmScale](float x) -> float
             {
                 if ( x <= 0.0f )
                 {
